requirements.txt
aniso8601==7.0.0
astroid==2.4.2
Django==2.1.4
django-environ==0.4.5
django-filter==2.0.0
django-graphql-jwt==0.2.0
graphene==2.1.8
graphene-django==2.2.0
graphql-core==2.3.2
graphql-relay==2.0.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
promise==2.3
PyJWT==1.7.1
pylint==2.6.0
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
toml==0.10.1
wrapt==1.12.1
